+++ bootique-di/src/main/java/io/bootique/di/spi/ProvidesHandler.java
@@ -2,14 +2,13 @@ package io.bootique.di.spi;
 
 import io.bootique.di.DIRuntimeException;
 import io.bootique.di.Key;
-import io.bootique.di.Module;
 import io.bootique.di.Scope;
 import io.bootique.di.TypeLiteral;
 
-import javax.inject.Named;
 import javax.inject.Provider;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
+import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -21,21 +20,20 @@ import java.util.Collections;
  */
 class ProvidesHandler {
 
-    private DefaultInjector injector;
-    private Class<? extends Annotation> providesAnnotation;
+    private final DefaultInjector injector;
+    private final Class<? extends Annotation> providesAnnotation;
 
     ProvidesHandler(DefaultInjector injector, Class<? extends Annotation> providesAnnotation) {
         this.injector = injector;
         this.providesAnnotation = providesAnnotation;
     }
 
-    Collection<KeyBindingPair<?>> bindingsFromAnnotatedMethods(Module module) {
+    Collection<KeyBindingPair<?>> bindingsFromAnnotatedMethods(Object module) {
 
         Collection<KeyBindingPair<?>> bindings = Collections.emptyList();
 
-        // consider annotated public methods in the module class and superclasses.
-
-        for (Method m : module.getClass().getMethods()) {
+        // consider annotated methods in the module class
+        for (Method m : module.getClass().getDeclaredMethods()) {
             Annotation pa = m.getAnnotation(providesAnnotation);
             if (pa != null) {
 
@@ -53,29 +51,31 @@ class ProvidesHandler {
         return bindings;
     }
 
-    private void validateProvidesMethod(Module module, Method method) {
+    private void validateProvidesMethod(Object module, Method method) {
 
         if (method.getReturnType().equals(Void.TYPE)) {
             throw new DIRuntimeException(
-                    "Invalid 'void' provider method '%s' on module '%s' is void. To be a proper provider method, it must return a value",
+                    "Invalid 'void' provider method '%s()' on module '%s' is void. To be a proper provider method, it must return a value",
                     method.getName(), module.getClass().getSimpleName());
         }
     }
 
-    private <T> KeyBindingPair<T> createBindingPair(Module module, Method method) {
+    private <T> KeyBindingPair<T> createBindingPair(Object module, Method method) {
 
-        Key<T> key = createKey(method.getGenericReturnType(), extractQualifier(method.getDeclaredAnnotations()));
+        Key<T> key = createKey(method.getGenericReturnType(), extractQualifier(method, method.getDeclaredAnnotations()));
         Binding<T> binding = createBinding(module, method);
 
         return new KeyBindingPair<>(key, binding);
     }
 
-    private Annotation extractQualifier(Annotation[] annotations) {
+    private Annotation extractQualifier(Method method, Annotation[] annotations) {
         Annotation found = null;
         for (Annotation a : annotations) {
-            if(DIUtil.isQualifyingAnnotation(a)) {
-                if(found != null) {
-                    throw new DIRuntimeException("Multiple qualifying annotations found");
+            if (DIUtil.isQualifyingAnnotation(a)) {
+                if (found != null) {
+                    throw new DIRuntimeException("Multiple qualifying annotations found for method '%s()' or its parameter on module '%s'"
+                            , method.getName()
+                            , method.getDeclaringClass().getSimpleName());
                 }
                 found = a;
             }
@@ -85,14 +85,26 @@ class ProvidesHandler {
     }
 
     private <T> Key<T> createKey(Type bindingType, Annotation qualifier) {
+        if (isProviderType(bindingType)) {
+            // Use provider generic argument as key
+            bindingType = ((ParameterizedType) bindingType).getActualTypeArguments()[0];
+        }
         return Key.get(TypeLiteral.of(bindingType), qualifier);
     }
 
-    private <T> Binding<T> createBinding(Module module, Method method) {
+    private boolean isProviderType(Type type) {
+        if (type instanceof ParameterizedType) {
+            ParameterizedType parameterizedType = (ParameterizedType) type;
+            return parameterizedType.getRawType().equals(Provider.class);
+        }
+        return false;
+    }
+
+    private <T> Binding<T> createBinding(Object module, Method method) {
         return new Binding<>(createProvider(module, method), createScope(method));
     }
 
-    private <T> Provider<T> createProvider(Module module, Method method) {
+    private <T> Provider<T> createProvider(Object module, Method method) {
 
         Provider<?>[] argumentProviders = createArgumentProviders(method);
 
@@ -106,12 +118,12 @@ class ProvidesHandler {
                 }
 
                 // supporting both 'static' and instance methods..
-                // TODO: accessibility - non-public inner and top-level classes... Does 'setAccessible' work across Java 9 modules?
+                method.setAccessible(true);
                 @SuppressWarnings("unchecked")
                 T result = (T) method.invoke(module, arguments);
                 return result;
             } catch (Exception e) {
-                throw new DIRuntimeException("Error invoking provider method '%s' on module '%s'",
+                throw new DIRuntimeException("Error invoking provider method '%s()' on module '%s'",
                         e,
                         method.getName(),
                         module.getClass().getSimpleName());
@@ -132,12 +144,16 @@ class ProvidesHandler {
         Provider<?>[] providers = new Provider[len];
 
         for (int i = 0; i < len; i++) {
-
-            Annotation qualifier = extractQualifier(paramAnnotations[i]);
+            Annotation qualifier = extractQualifier(method, paramAnnotations[i]);
             Key<?> key = createKey(params[i], qualifier);
 
-            // resolve the actual provider lazily
-            providers[i] = () -> injector.getProvider(key).get();
+            if (isProviderType(params[i])) {
+                // will resolve to provider of provider
+                providers[i] = () -> injector.getProvider(key);
+            } else {
+                // resolve the actual provider lazily
+                providers[i] = () -> injector.getProvider(key).get();
+            }
         }
 
         return providers;