+++ bootique-di/src/test/java/io/bootique/di/ProvidesIT.java
@@ -1,8 +1,13 @@
 package io.bootique.di;
 
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
 import org.junit.Test;
 
 import javax.inject.Named;
+import javax.inject.Provider;
+import javax.inject.Qualifier;
 
 import static org.junit.Assert.assertEquals;
 
@@ -25,6 +30,19 @@ public class ProvidesIT {
     }
 
     @Test
+    public void testProvides_Standalone_AnonymousClass() {
+        Injector injector = DIBootstrap.createInjector(new BaseModule() {
+            @Provides
+            Service1 createService() {
+                return () -> "provideService1";
+            }
+        });
+
+        Service1 s1 = injector.getInstance(Service1.class);
+        assertEquals("provideService1", s1.doIt());
+    }
+
+    @Test
     public void testProvides_Standalone_Named() {
         Injector injector = DIBootstrap.createInjector(new TestModule_NamedService());
 
@@ -41,8 +59,24 @@ public class ProvidesIT {
     }
 
     @Test
-    public void testProvides_Chain_NamedPartameter() {
+    public void testProvides_Chain_NamedParameter() {
         Injector injector = DIBootstrap.createInjector(new TestModule_NamedParameter());
+
+        Service2 s2 = injector.getInstance(Service2.class);
+        assertEquals("provideService2_provideService1", s2.doIt());
+    }
+
+    @Test
+    public void testProvides_Chain_ProviderParameter() {
+        Injector injector = DIBootstrap.createInjector(new TestModule_ServiceChain_ProviderParameter());
+
+        Service2 s2 = injector.getInstance(Service2.class);
+        assertEquals("provideService2_provideService1", s2.doIt());
+    }
+
+    @Test
+    public void testProvides_Chain_QualifiedProviderParameter() {
+        Injector injector = DIBootstrap.createInjector(new TestModule_QualifiedProviderParameter());
 
         Service2 s2 = injector.getInstance(Service2.class);
         assertEquals("provideService2_provideService1", s2.doIt());
@@ -51,6 +85,11 @@ public class ProvidesIT {
     @Test(expected = DIRuntimeException.class)
     public void testProvides_Invalid() {
         DIBootstrap.createInjector(new TestModule_InvalidProvider());
+    }
+
+    @Test(expected = DIRuntimeException.class)
+    public void testProvides_InvalidQualifier() {
+        DIBootstrap.createInjector(new TestModule_InvalidQualifier());
     }
 
     interface Service1 {
@@ -85,8 +124,21 @@ public class ProvidesIT {
         }
 
         @Provides
-        public static Service2 provideService2(Service1 s1) {
+        static Service2 provideService2(Service1 s1) {
             return () -> "provideService2_" + s1.doIt();
+        }
+    }
+
+    public static class TestModule_ServiceChain_ProviderParameter extends BaseModule {
+
+        @Provides
+        public static Service1 provideService1() {
+            return () -> "provideService1";
+        }
+
+        @Provides
+        static Service2 provideService2(Provider<Service1> s1) {
+            return () -> "provideService2_" + s1.get().doIt();
         }
     }
 
@@ -94,6 +146,16 @@ public class ProvidesIT {
 
         @Provides
         public void invalidProvides() {
+        }
+    }
+
+    public static class TestModule_InvalidQualifier extends BaseModule {
+
+        @TestQualifier
+        @Named("s1")
+        @Provides
+        public Service1 invalidProvides() {
+            return () -> "provideService1";
         }
     }
 
@@ -108,6 +170,11 @@ public class ProvidesIT {
 
     public static class TestModule_NamedParameter extends BaseModule {
 
+        @Provides
+        public static Service1 provideUnnamedService1() {
+            return () -> "provideService1_unnamed";
+        }
+
         @Named("s1")
         @Provides
         public static Service1 provideService1() {
@@ -115,9 +182,33 @@ public class ProvidesIT {
         }
 
         @Provides
-        public static Service2 provideService2(@Named("s1") Service1 s1) {
+        public Service2 provideService2(@Named("s1") Service1 s1) {
             return () -> "provideService2_" + s1.doIt();
         }
+    }
+
+    public static class TestModule_QualifiedProviderParameter extends BaseModule {
+
+        @Provides
+        public static Service1 provideUnnamedService1() {
+            return () -> "provideService1_unqualified";
+        }
+
+        @TestQualifier
+        @Provides
+        public static Service1 provideService1() {
+            return () -> "provideService1";
+        }
+
+        @Provides
+        public Service2 provideService2(@TestQualifier Provider<Service1> s1) {
+            return () -> "provideService2_" + s1.get().doIt();
+        }
+    }
+
+    @Qualifier
+    @Retention(RetentionPolicy.RUNTIME)
+    @interface TestQualifier {
     }
 }